# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A self-contained YouTube Music player backend that authenticates via browser cookies,\nresolves audio streams through the Innertube internal API, and plays audio locally.\nDesigned as a modular music player daemon for TUIs, Raycast extensions, and other clients.",
        "title": "YouTube Music API",
        "contact": {
            "name": "YTMusic API"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/player/next": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Advances to the next track in the queue. Returns the new player state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Skip to next track",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player/pause": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Pauses the current track if playing, or resumes if paused.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Toggle pause/resume",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player/play": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Resolves the audio stream for a YouTube Music video ID and begins playback.\nIf the track is not already in the queue, it is added at the current position.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Play a track",
                "parameters": [
                    {
                        "description": "Track to play",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PlayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player/previous": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Goes back to the previous track in the queue. Returns the new player state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Go to previous track",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player/repeat": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sets repeat mode: \"off\", \"all\", \"one\", or \"cycle\" (cycles to next mode).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Set repeat mode",
                "parameters": [
                    {
                        "description": "Repeat mode",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RepeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player/shuffle": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Toggles shuffle mode on or off. When enabled, queue is shuffled.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Toggle shuffle mode",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player/state": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the current playback state including current track, pause status, and queue info.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Get player state",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player/stop": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Stops the current track and clears the playing state. Queue is preserved.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Stop playback",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player/volume": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sets the playback volume (0-100).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "Set playback volume",
                "parameters": [
                    {
                        "description": "Volume level",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VolumeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerState"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AlbumInfo": {
            "type": "object",
            "properties": {
                "browse_id": {
                    "type": "string",
                    "example": "MPREb_..."
                },
                "name": {
                    "type": "string",
                    "example": "Whenever You Need Somebody"
                }
            }
        },
...
        "models.Artist": {
            "type": "object",
            "properties": {
                "browse_id": {
                    "type": "string",
                    "example": "UCuAXFkgsw1L7xaCfnd5JJOw"
                },
                "name": {
                    "type": "string",
                    "example": "Rick Astley"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
...
        "models.PlayRequest": {
            "type": "object",
            "required": [
                "video_id"
            ],
            "properties": {
                "video_id": {
                    "type": "string",
                    "example": "dQw4w9WgXcQ"
                }
            }
        },
        "models.PlayerState": {
            "type": "object",
            "properties": {
                "current_track": {
                    "$ref": "#/definitions/models.Track"
                },
                "is_paused": {
                    "type": "boolean",
                    "example": false
                },
                "is_playing": {
                    "type": "boolean",
                    "example": true
                },
                "queue_length": {
                    "type": "integer",
                    "example": 5
                },
                "queue_position": {
                    "type": "integer",
                    "example": 0
                },
                "repeat": {
                    "description": "\"off\", \"all\", \"one\"",
                    "type": "string",
                    "example": "off"
                },
                "shuffle": {
                    "type": "boolean",
                    "example": false
                },
                "volume": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "models.RepeatRequest": {
            "type": "object",
            "required": [
                "repeat"
            ],
            "properties": {
                "repeat": {
                    "description": "\"off\", \"all\", \"one\"",
                    "type": "string",
                    "example": "all"
                }
            }
        },
...
        "models.Track": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string",
                    "example": "Whenever You Need Somebody"
                },
                "album_info": {
                    "$ref": "#/definitions/models.AlbumInfo"
                },
                "artist": {
                    "type": "string",
                    "example": "Rick Astley"
                },
                "artists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Artist"
                    }
                },
                "duration": {
                    "type": "string",
                    "example": "3:33"
                },
                "duration_ms": {
                    "type": "integer",
                    "example": 213000
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://i.ytimg.com/vi/dQw4w9WgXcQ/hqdefault.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Never Gonna Give You Up"
                },
                "video_id": {
                    "type": "string",
                    "example": "dQw4w9WgXcQ"
                }
            }
        },
...
        "models.VolumeRequest": {
            "type": "object",
            "required": [
                "volume"
            ],
            "properties": {
                "volume": {
                    "type": "integer",
                    "example": 80
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Session-Token",
            "in": "header"
        }
    }
}